+++ backend/users/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
+from django.contrib.auth.models import Group
 
 from users.models import User
 
@@ -7,19 +8,8 @@ from users.models import User
 @admin.register(User)
 class UserAdmin(UserAdmin):
     model = User
-    list_display = (
-        "id",
-        "username",
-        "email",
-        "is_superuser",
-        "is_staff",
-        "is_active",
-    )
-    list_filter = (
-        "username",
-        "is_staff",
-        "is_active",
-    )
+    list_display = ("id", "username", "email", "is_superuser", "is_staff", "is_active")
+    list_filter = ("username", "is_staff", "is_active")
     fieldsets = (
         ("Profile", {"fields": ("username", "email", "password")}),
         ("Permissions", {"fields": ("is_superuser", "is_staff", "is_active")}),
@@ -42,3 +32,6 @@ class UserAdmin(UserAdmin):
     )
     search_fields = ("username",)
     ordering = ("username",)
+
+
+admin.site.unregister(Group)