+++ CHANGELOG.md
@@ -2,4 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.2](https://github.com/Recedivies/Coffeengo/compare/v0.0.1...v0.0.2) (2022-08-08)
+
+
+### Features
+
+* added auth axios fe ([ed50a6d](https://github.com/Recedivies/Coffeengo/commit/ed50a6d32d307be4708dd168b52b43bce58524b9))
+* added fe routing ([559b9da](https://github.com/Recedivies/Coffeengo/commit/559b9dad65c014a55e6f7334503d9c8e2fc9939d))
+* added hooks context ([b621825](https://github.com/Recedivies/Coffeengo/commit/b621825db3886b7648d668b793c08214a63956bb))
+* added layout fe ([9079761](https://github.com/Recedivies/Coffeengo/commit/9079761b396218b1d9100769f2f3161ffe1d07ea))
+* added lib clsx for className ([493a3d8](https://github.com/Recedivies/Coffeengo/commit/493a3d8ffe6295be03864205469106ce33ddf32f))
+* added Login auth ([204d089](https://github.com/Recedivies/Coffeengo/commit/204d0897f7835d9a53c3ea59e11318430ad308dd))
+* added pages components ([20331e0](https://github.com/Recedivies/Coffeengo/commit/20331e0274d320187712b10463a67ab80d89833d))
+* added Register auth ([17170ec](https://github.com/Recedivies/Coffeengo/commit/17170ec1ed064ea75ecbfac1dd3553233e622713))
+* added types in fe ([ab4c8ab](https://github.com/Recedivies/Coffeengo/commit/ab4c8abd94a33ac4a316a91321bf694b514ff124))
+* added users auth ([6ff8939](https://github.com/Recedivies/Coffeengo/commit/6ff8939916e55fa7bc628d7796a4c6e1c04f7848))
+* added users models ([181d7ec](https://github.com/Recedivies/Coffeengo/commit/181d7ec509765446ba52112630a6b4491fd99751))
+* added utils components ([5036103](https://github.com/Recedivies/Coffeengo/commit/503610387ed98168601c05d249e351ae824fc590))
+* init tailwind css ([8502cc3](https://github.com/Recedivies/Coffeengo/commit/8502cc3c79d715c41e28c8ba922d09dcb0a1c635))
+
+
+### Bug Fixes
+
+* linting ([3280148](https://github.com/Recedivies/Coffeengo/commit/3280148242f940533b4cf52d1d374f975f39525d))
+* tests be and url path for pass change ([f75114a](https://github.com/Recedivies/Coffeengo/commit/f75114a6018a70ec46f8a7ae43700e3d49dd8fe9))
+* update endpoint for password change ([07a5db2](https://github.com/Recedivies/Coffeengo/commit/07a5db219efd0706bfcc0b550e7d574b4ea426a9))
+* update stuff and docs ([14b52e1](https://github.com/Recedivies/Coffeengo/commit/14b52e129800b6197317539cf22942b00e36f832))
+
 ### 0.0.1 (2022-07-13)